requirements.txt
flask
python-dotenv
line-bot-sdk
gunicorn
